+++ stm32/cores/arduino/Klangstrom/NodeMixer4Stereo.hpp
@@ -88,12 +88,14 @@ namespace klang {
         }
                 
         void update(CHANNEL_ID pChannel, SIGNAL_TYPE* pAudioBlock) {
+            const bool b0 = (mConnection_CH_IN_SIGNAL_0!=nullptr);
+            const bool b1 = (mConnection_CH_IN_SIGNAL_1!=nullptr);
+            const bool b2 = (mConnection_CH_IN_SIGNAL_2!=nullptr);
+            const bool b3 = (mConnection_CH_IN_SIGNAL_3!=nullptr);
+            const uint8_t mSignalInputCounter = b0 + b1 + b2 + b3;
             if (is_not_updated()
                 && pChannel == CH_OUT_SIGNAL
-                &&(mConnection_CH_IN_SIGNAL_0!=nullptr 
-                    && mConnection_CH_IN_SIGNAL_1!=nullptr
-                    && mConnection_CH_IN_SIGNAL_2!=nullptr
-                    && mConnection_CH_IN_SIGNAL_3!=nullptr)) {
+                && mSignalInputCounter > 0) {
                 AUDIO_BLOCK_ID mBlock_SIGNAL_0  = AudioBlockPool::NO_ID;
                 AUDIO_BLOCK_ID mBlock_SIGNAL_1  = AudioBlockPool::NO_ID;
                 AUDIO_BLOCK_ID mBlock_SIGNAL_2  = AudioBlockPool::NO_ID;
@@ -103,22 +105,22 @@ namespace klang {
                 SIGNAL_TYPE* mBlockData_SIGNAL_2 = nullptr;
                 SIGNAL_TYPE* mBlockData_SIGNAL_3 = nullptr;
                 
-                if (mConnection_CH_IN_SIGNAL_0 != nullptr) {
+                if (b0) {
                     mBlock_SIGNAL_0 = AudioBlockPool::instance().request();
                     mBlockData_SIGNAL_0 = AudioBlockPool::instance().data(mBlock_SIGNAL_0);
                     mConnection_CH_IN_SIGNAL_0->update(mBlock_SIGNAL_0);
                 }
-                if (mConnection_CH_IN_SIGNAL_1 != nullptr) {
+                if (b1) {
                     mBlock_SIGNAL_1 = AudioBlockPool::instance().request();
                     mBlockData_SIGNAL_1 = AudioBlockPool::instance().data(mBlock_SIGNAL_1);
                     mConnection_CH_IN_SIGNAL_1->update(mBlock_SIGNAL_1);
                 }
-                if (mConnection_CH_IN_SIGNAL_2 != nullptr) {
+                if (b2) {
                     mBlock_SIGNAL_2 = AudioBlockPool::instance().request();
                     mBlockData_SIGNAL_2 = AudioBlockPool::instance().data(mBlock_SIGNAL_2);
                     mConnection_CH_IN_SIGNAL_2->update(mBlock_SIGNAL_2);
                 }
-                if (mConnection_CH_IN_SIGNAL_3 != nullptr) {
+                if (b3) {
                     mBlock_SIGNAL_3 = AudioBlockPool::instance().request();
                     mBlockData_SIGNAL_3 = AudioBlockPool::instance().data(mBlock_SIGNAL_3);
                     mConnection_CH_IN_SIGNAL_3->update(mBlock_SIGNAL_3);
@@ -129,22 +131,20 @@ namespace klang {
                 SIGNAL_TYPE* mBlockData_CH_OUT_SIGNAL_LEFT = AudioBlockPool::instance().data(mBlock_CH_OUT_SIGNAL_LEFT);
                 SIGNAL_TYPE* mBlockData_CH_OUT_SIGNAL_RIGHT = AudioBlockPool::instance().data(mBlock_CH_OUT_SIGNAL_RIGHT);
 
-                const bool b0 = (mBlockData_SIGNAL_0!=nullptr);
-                const bool b1 = (mBlockData_SIGNAL_1!=nullptr);
-                const bool b2 = (mBlockData_SIGNAL_2!=nullptr);
-                const bool b3 = (mBlockData_SIGNAL_3!=nullptr);
-                const float pL0 = mPan[0];
-                const float pR0 = 1.0 - mPan[0];
-                const float pL1 = mPan[1];
-                const float pR1 = 1.0 - mPan[1];
-                const float pL2 = mPan[2];
-                const float pR2 = 1.0 - mPan[2];
-                const float pL3 = mPan[3];
-                const float pR3 = 1.0 - mPan[3];
-                const uint8_t mSignalInputCounter = (b0 ? 1 : 0) +
-                                                    (b1 ? 1 : 0) +
-                                                    (b2 ? 1 : 0) +
-                                                    (b3 ? 1 : 0);
+                // @todo(this should range from [-1.0...1.0] and not from [0.0...1.0])
+                const float mPan0 = mPan[SIGNAL_CHANNEL::SIGNAL_0] * 0.5 + 0.5;
+                const float mPan1 = mPan[SIGNAL_CHANNEL::SIGNAL_1] * 0.5 + 0.5;
+                const float mPan2 = mPan[SIGNAL_CHANNEL::SIGNAL_2] * 0.5 + 0.5;
+                const float mPan3 = mPan[SIGNAL_CHANNEL::SIGNAL_3] * 0.5 + 0.5;
+                const float pR0 = mPan0;
+                const float pL0 = 1.0 - mPan0;
+                const float pR1 = mPan1;
+                const float pL1 = 1.0 - mPan1;
+                const float pR2 = mPan2;
+                const float pL2 = 1.0 - mPan2;
+                const float pR3 = mPan3;
+                const float pL3 = 1.0 - mPan3;
+
                 const float mRatio = 1.0 / mSignalInputCounter;
                 for (uint16_t i=0; i < KLANG_SAMPLES_PER_AUDIO_BLOCK; ++i) {
                     float sL = 0.0;
@@ -191,34 +191,46 @@ namespace klang {
                 flag_updated();
             }
             if (pChannel == CH_OUT_SIGNAL_LEFT) {
-                SIGNAL_TYPE* mBlockData_CH_OUT_SIGNAL_LEFT = AudioBlockPool::instance().data(mBlock_CH_OUT_SIGNAL_LEFT);
+                if (mBlock_CH_OUT_SIGNAL_LEFT != AudioBlockPool::NO_ID) {
+                    SIGNAL_TYPE* mBlockData_CH_OUT_SIGNAL_LEFT = AudioBlockPool::instance().data(mBlock_CH_OUT_SIGNAL_LEFT);
 #ifdef KLST_USE_DSP
-                arm_copy_f32(mBlockData_CH_OUT_SIGNAL_LEFT,pAudioBlock,KLANG_SAMPLES_PER_AUDIO_BLOCK);
+                    arm_copy_f32(mBlockData_CH_OUT_SIGNAL_LEFT,pAudioBlock,KLANG_SAMPLES_PER_AUDIO_BLOCK);
 #else
-                memcpy( pAudioBlock,
-                        mBlockData_CH_OUT_SIGNAL_LEFT,
-                        sizeof(SIGNAL_TYPE) * KLANG_SAMPLES_PER_AUDIO_BLOCK);
+                    memcpy( pAudioBlock,
+                            mBlockData_CH_OUT_SIGNAL_LEFT,
+                            sizeof(SIGNAL_TYPE) * KLANG_SAMPLES_PER_AUDIO_BLOCK);
 #endif
+                }
             } else if (pChannel == CH_OUT_SIGNAL_RIGHT) {
-                SIGNAL_TYPE* mBlockData_CH_OUT_SIGNAL_RIGHT = AudioBlockPool::instance().data(mBlock_CH_OUT_SIGNAL_RIGHT);
+                if (mBlock_CH_OUT_SIGNAL_RIGHT != AudioBlockPool::NO_ID) {
+                    SIGNAL_TYPE* mBlockData_CH_OUT_SIGNAL_RIGHT = AudioBlockPool::instance().data(mBlock_CH_OUT_SIGNAL_RIGHT);
 #ifdef KLST_USE_DSP
-                arm_copy_f32(mBlockData_CH_OUT_SIGNAL_LEFT,pAudioBlock,KLANG_SAMPLES_PER_AUDIO_BLOCK);
+                    arm_copy_f32(mBlockData_CH_OUT_SIGNAL_LEFT,pAudioBlock,KLANG_SAMPLES_PER_AUDIO_BLOCK);
 #else
-                memcpy( pAudioBlock,
-                        mBlockData_CH_OUT_SIGNAL_RIGHT,
-                        sizeof(SIGNAL_TYPE) * KLANG_SAMPLES_PER_AUDIO_BLOCK);
+                    memcpy( pAudioBlock,
+                            mBlockData_CH_OUT_SIGNAL_RIGHT,
+                            sizeof(SIGNAL_TYPE) * KLANG_SAMPLES_PER_AUDIO_BLOCK);
 #endif
+                }
             }
         }
         
         void set_mix(SIGNAL_CHANNEL pChannel, SIGNAL_TYPE pValue) {
             mMix[pChannel] = pValue;
         }
-        
+
+        SIGNAL_TYPE get_mix(uint8_t pChannel) {
+            return mMix[pChannel];
+        }
+                
         void set_pan(SIGNAL_CHANNEL pChannel, SIGNAL_TYPE pValue) {
             mPan[pChannel] = pValue;
         }
         
+        SIGNAL_TYPE get_pan(uint8_t pChannel) {
+            return mPan[pChannel];
+        }
+
         void set_command(KLANG_CMD_TYPE pCommand, KLANG_CMD_TYPE* pPayLoad) {
             switch (pCommand) {
                 case KLANG_SET_MIX_F32:
@@ -240,7 +252,7 @@ namespace klang {
         Connection* mConnection_CH_IN_SIGNAL_3  = nullptr;
         static const uint8_t NUM_SIGNAL_INPUT_CHANNELS = 4;
         SIGNAL_TYPE mMix[NUM_SIGNAL_INPUT_CHANNELS] = {1.0, 1.0, 1.0, 1.0};
-        SIGNAL_TYPE mPan[NUM_SIGNAL_INPUT_CHANNELS] = {0.5, 0.5, 0.5, 0.5};
+        SIGNAL_TYPE mPan[NUM_SIGNAL_INPUT_CHANNELS] = {0.0, 0.0, 0.0, 0.0};
     };
 }